% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{tpl.accepted}
\alias{tpl.accepted}
\alias{tpl.accepted.index}
\alias{tpl.misapplied}
\alias{tpl.names}
\alias{tpl.synonyms,}
\alias{tpl.unresolved,}
\alias{words,}
\title{The Plant List}
\format{Lists and data frames}
\description{
A few datasets containing data scraped from theplantlist.com
}

